Change the header in the README
Change the header in the README to replace the information that was removed from the repository’s name. The repository’s name was too long to fit on one line in my resume so ‘Health’ and ‘Cost’ had to be removed.  Add ‘Cost’ to the README header to clarify what the project is about.